Abstract Affiliations edit
